Add spec for audit entries
- As the audit capabilities have become more and more specific, it makes sense to create its own spec so we can declutter the remaining ones and easily check all the logic for the audit in one spec
