+++ spec/system/planning_applications/red_line_boundary_change_validation_request_spec.rb
@@ -83,21 +83,6 @@ RSpec.describe "Requesting map changes to a planning application", type: :system
     expect(page).to have_content("Provide a reason for changes")
   end
 
-  it "displays the details of the received request in the audit log" do
-    create :audit, planning_application_id: planning_application.id, activity_type: "red_line_boundary_change_validation_request_received", activity_information: 1, audit_comment: { response: "rejected", reason: "The boundary was too small" }.to_json, api_user: api_user
-
-    sign_in assessor
-    visit planning_application_path(planning_application)
-
-    click_button "Key application dates"
-    click_link "Activity log"
-
-    expect(page).to have_text("Received: request for change (red line boundary#1)")
-    expect(page).to have_text("The boundary was too small")
-    expect(page).to have_text("rejected")
-    expect(page).to have_text("Applicant / Agent via Api Wizard")
-  end
-
   it "updates the notified_at date of an open request when application is invalidated" do
     new_planning_application = create :planning_application, :not_started, local_authority: @default_local_authority
     request = create :red_line_boundary_change_validation_request, planning_application: new_planning_application, state: "open", created_at: 12.days.ago
