+++ spec/system/planning_applications/replacement_document_validation_request_spec.rb
@@ -82,20 +82,6 @@ RSpec.describe "Requesting document changes to a planning application", type: :s
     expect(page).not_to have_content(invalid_document.name.to_s)
   end
 
-  it "displays the details of the received request in the audit log" do
-    create :audit, planning_application_id: planning_application.id, activity_type: "replacement_document_validation_request_received", activity_information: 1, audit_comment: "floor_plan.pdf", api_user: api_user
-
-    sign_in assessor
-    visit planning_application_path(planning_application)
-
-    click_button "Key application dates"
-    click_link "Activity log"
-
-    expect(page).to have_text("Received: request for change (replacement document#1)")
-    expect(page).to have_text("floor_plan.pdf")
-    expect(page).to have_text("Applicant / Agent via Api Wizard")
-  end
-
   context "Invalidation updates replacement document validation request" do
     it "updates the notified_at date of an open request when application is invalidated" do
       new_planning_application = create :planning_application, :not_started, local_authority: @default_local_authority
