+++ spec/system/planning_applications/update_spec.rb
@@ -110,62 +110,4 @@ RSpec.describe "Updating a planning application", type: :system do
 
     expect(page).to have_text("293844848")
   end
-
-  it "creates an audit entry for every update action" do
-    fill_in "Description", with: "doing more than great things, doing wonderful things."
-    fill_in "Day", with: "3"
-    fill_in "Month", with: "10"
-    fill_in "Year", with: "2021"
-    fill_in "Address 1", with: "2 Streatham High Road"
-    fill_in "Address 2", with: "Streatham"
-    fill_in "Town", with: "Crystal Palace"
-    fill_in "County", with: "London"
-    fill_in "Postcode", with: "SW16 1DB"
-    fill_in "UPRN", with: "294884040"
-    within "form", text: "Has the work been started?" do
-      choose "Yes"
-    end
-    within ".applicant-information" do
-      fill_in "First name", with: "Pearly"
-      fill_in "Last name", with: "Poorly"
-      fill_in "Email address", with: "pearly@poorly.com"
-      fill_in "UK telephone number", with: "0777773949494312"
-    end
-    within ".agent-information" do
-      fill_in "First name", with: "Agentina"
-      fill_in "Last name", with: "Agentino"
-      fill_in "Email address", with: "agentina@agentino.com"
-      fill_in "UK telephone number", with: "923838484492939"
-    end
-
-    click_button "Save"
-    click_button "Key application dates"
-    click_link "Activity log"
-
-    expect(page).to have_text("Description updated")
-    expect(page).to have_text("Changed from: Back shack")
-    expect(page).to have_text("Changed to: doing more than great things, doing wonderful things.")
-
-    expect(page).to have_text("Address 1 updated")
-    expect(page).to have_text("Changed from: 1 golden street")
-    expect(page).to have_text("Changed to: 2 Streatham High Road")
-
-    expect(page).to have_text("Address 2 updated")
-    expect(page).to have_text("Changed from: Southwark")
-    expect(page).to have_text("Changed to: Streatham")
-
-    expect(page).to have_text("Applicant first name updated")
-    expect(page).to have_text("Applicant last name updated")
-    expect(page).to have_text("Applicant phone updated")
-    expect(page).to have_text("Applicant email updated")
-    expect(page).to have_text("Agent first name updated")
-    expect(page).to have_text("Agent last name updated")
-    expect(page).to have_text("Agent phone updated")
-    expect(page).to have_text("Agent email updated")
-    expect(page).to have_text("County updated")
-    expect(page).to have_text("Postcode updated")
-    expect(page).to have_text("Town updated")
-    expect(page).to have_text("Work status updated")
-    expect(page).to have_text("Payment reference updated")
-  end
 end
